# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Covalence API",
        "description": "Hybrid GraphRAG knowledge engine API",
        "license": {
            "name": "MIT",
            "identifier": "MIT"
        },
        "version": "0.1.0"
    },
    "paths": {
...
        "/mcp/tools/call": {
            "post": {
                "tags": [
                    "mcp"
                ],
                "summary": "Call an MCP tool by name.",
                "operationId": "call_tool",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/McpToolCall"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Tool result",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/McpToolResult"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/nodes/{id}": {
            "get": {
                "tags": [
                    "nodes"
                ],
                "summary": "Get a node by ID.",
                "operationId": "get_node",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Node ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Node found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NodeResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Node not found"
                    }
                }
            }
        },
...
        "/nodes/{id}/provenance": {
            "get": {
                "tags": [
                    "nodes"
                ],
                "summary": "Get the provenance chain for a node.",
                "operationId": "get_provenance",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Node ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Provenance chain",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProvenanceResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "McpContent": {
                "type": "object",
                "description": "A single content block in an MCP result.",
                "required": [
                    "type",
                    "text"
                ],
                "properties": {
                    "text": {
                        "type": "string",
                        "description": "Text payload (usually JSON-serialized data)."
                    },
                    "type": {
                        "type": "string",
                        "description": "Content type (typically \"text\")."
                    }
                }
            },
...
            "McpToolCall": {
                "type": "object",
                "description": "Request to call an MCP tool.",
                "required": [
                    "name",
                    "arguments"
                ],
                "properties": {
                    "arguments": {
                        "description": "Arguments as a JSON object."
                    },
                    "name": {
                        "type": "string",
                        "description": "Name of the tool to invoke."
                    }
                }
            },
            "McpToolResult": {
                "type": "object",
                "description": "Result from an MCP tool call.",
                "required": [
                    "content",
                    "is_error"
                ],
                "properties": {
                    "content": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/McpContent"
                        },
                        "description": "Content blocks in the result."
                    },
                    "is_error": {
                        "type": "boolean",
                        "description": "Whether the tool call resulted in an error."
                    }
                }
            },
...
            "NodeResponse": {
                "type": "object",
                "description": "Response for a node entity.",
                "required": [
                    "id",
                    "canonical_name",
                    "node_type",
                    "properties",
                    "clearance_level",
                    "first_seen",
                    "last_seen",
                    "mention_count"
                ],
                "properties": {
                    "canonical_name": {
                        "type": "string"
                    },
                    "clearance_level": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "description": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "first_seen": {
                        "type": "string"
                    },
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "last_seen": {
                        "type": "string"
                    },
                    "mention_count": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "node_type": {
                        "type": "string"
                    },
                    "properties": {}
                }
            },
...
            "ProvenanceResponse": {
                "type": "object",
                "description": "Provenance chain response.",
                "required": [
                    "node_id",
                    "extraction_count",
                    "chunk_count",
                    "source_count"
                ],
                "properties": {
                    "chunk_count": {
                        "type": "integer",
                        "description": "Number of source chunks.",
                        "minimum": 0
                    },
                    "extraction_count": {
                        "type": "integer",
                        "description": "Number of extraction records.",
                        "minimum": 0
                    },
                    "node_id": {
                        "type": "string",
                        "format": "uuid",
                        "description": "Node ID."
                    },
                    "source_count": {
                        "type": "integer",
                        "description": "Number of originating sources.",
                        "minimum": 0
                    }
                }
            },
...
        }
    },
...
}